Authorization added to routs
