+++ package.json
@@ -3,6 +3,7 @@
   "main": "server.js",
   "dependencies": {
     "@okta/jwt-verifier": "0.0.14",
+    "axios": "^0.18.0",
     "body-parser": "^1.18.3",
     "cors": "^2.8.4",
     "dotenv": "^6.1.0",
